+++ HvZ-TheGame/HvZ-TheGame/Character.h
@@ -6,9 +6,9 @@ class Character :public BaseClass {
 
 public:
 	sf::Vector3i gridDestination;
-	sf::Vector3i pointDestination;
+	sf::Vector3f pointDestination;
 
-	Character(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, sf::Vector3i pointLocation);
+	Character(World& world, const sf::Texture& texture, sf::Vector3i gridLocation, sf::Vector3f pointLocation);
 	~Character();
 
 	virtual	void move(int x, int y, int z);
